+++ R/add_posterior_coefficients.expandingwindow.R
@@ -4,6 +4,8 @@
 #' 
 #' @param object an object of class 'expandingwindow'.
 #' @param ... further arguments passed to or from other methods.
+#' @inheritParams add_posterior_coefficients.modellist
+#' @inheritSection add_posterior_coefficients.modellist Parallel simulation
 #' 
 #' @return The object in \code{object} with posterior draws added to each of its models, as
 #' described in \code{\link{add_posterior_coefficients.bvarmodel}}.
@@ -35,7 +37,12 @@
 #' model <- add_posterior_coefficients(model)
 #' 
 #' @export
-add_posterior_coefficients.expandingwindow <- function(object, ...){
+add_posterior_coefficients.expandingwindow <- function(object, ..., cores = 1){
+
+  if (.use_cluster(object, cores)) {
+    return(.simulate_models_in_parallel(object, add_posterior_coefficients, cores, ...,
+                                        seed = TRUE))
+  }
 
   object <- lapply(object, add_posterior_coefficients, ...)