+++ R/add_posterior_loglik.modellist.R
@@ -5,6 +5,8 @@
 #' @param object a list of class 'modellist', usually, the result of a call to
 #' \code{\link{add_posterior_coefficients}}.
 #' @param ... further arguments passed to or from other methods.
+#' @inheritParams add_posterior_coefficients.modellist
+#' @inheritSection add_posterior_coefficients.modellist Parallel simulation
 #' 
 #' @return The object in \code{object} with log-likelihood draws added to each of its models, as
 #' described in \code{\link{add_posterior_loglik.bvarmodel}}.
@@ -35,7 +37,11 @@
 #' object <- add_posterior_loglik(object)
 #' 
 #' @export
-add_posterior_loglik.modellist <- function(object, ...){
+add_posterior_loglik.modellist <- function(object, ..., cores = 1){
+
+  if (.use_cluster(object, cores)) {
+    return(.simulate_models_in_parallel(object, add_posterior_loglik, cores, ...))
+  }
   
   object <- lapply(object, add_posterior_loglik, ...)
   class(object) <- list("modellist", "list")